+++ lib/cache/memory.go
@@ -9,6 +9,7 @@ import (
 	"github.com/Jeffail/benthos/v3/lib/log"
 	"github.com/Jeffail/benthos/v3/lib/metrics"
 	"github.com/Jeffail/benthos/v3/lib/types"
+	"github.com/OneOfOne/xxhash"
 )
 
 //------------------------------------------------------------------------------
@@ -41,6 +42,7 @@ TTL is respected as usual.`,
 		FieldSpecs: docs.FieldSpecs{
 			docs.FieldCommon("ttl", "The TTL of each item in seconds. After this period an item will be eligible for removal during the next compaction."),
 			docs.FieldCommon("compaction_interval", "The period of time to wait before each compaction, at which point expired items are removed."),
+			docs.FieldAdvanced("shards", "A number of logical shards to spread keys across, increasing the shards can have a performance benefit when processing a large number of keys."),
 			docs.FieldCommon(
 				"init_values", "A table of key/value pairs that should be present in the cache on initialization. This can be used to create static lookup tables.",
 				map[string]string{
@@ -60,6 +62,7 @@ type MemoryConfig struct {
 	TTL                int               `json:"ttl" yaml:"ttl"`
 	CompactionInterval string            `json:"compaction_interval" yaml:"compaction_interval"`
 	InitValues         map[string]string `json:"init_values" yaml:"init_values"`
+	Shards             int               `json:"shards" yaml:"shards"`
 }
 
 // NewMemoryConfig creates a MemoryConfig populated with default values.
@@ -68,6 +71,7 @@ func NewMemoryConfig() MemoryConfig {
 		TTL:                300, // 5 Mins
 		CompactionInterval: "60s",
 		InitValues:         map[string]string{},
+		Shards:             1,
 	}
 }
 
@@ -78,18 +82,43 @@ type item struct {
 	ts    time.Time
 }
 
-// Memory is a memory based cache implementation.
-type Memory struct {
-	items          map[string]item
-	ttl            time.Duration
+type shard struct {
+	items map[string]item
+	ttl   time.Duration
+
 	compInterval   time.Duration
 	lastCompaction time.Time
 
-	stats        metrics.Type
-	mCompactions metrics.StatCounter
 	mKeys        metrics.StatGauge
+	mCompactions metrics.StatCounter
 
 	sync.RWMutex
+}
+
+func (s *shard) compaction() {
+	if s.compInterval == 0 {
+		return
+	}
+	if time.Since(s.lastCompaction) < s.compInterval {
+		return
+	}
+	s.mCompactions.Incr(1)
+	for k, v := range s.items {
+		if v.ts.IsZero() {
+			continue
+		}
+		if time.Since(v.ts) >= s.ttl {
+			delete(s.items, k)
+		}
+	}
+	s.lastCompaction = time.Now()
+}
+
+//------------------------------------------------------------------------------
+
+// Memory is a memory based cache implementation.
+type Memory struct {
+	shards []*shard
 }
 
 // NewMemory creates a new Memory cache type.
@@ -101,51 +130,67 @@ func NewMemory(conf Config, mgr types.Manager, log log.Modular, stats metrics.Ty
 			return nil, fmt.Errorf("failed to parse compaction interval string: %v", err)
 		}
 	}
-	items := map[string]item{}
+
+	m := &Memory{}
+	if conf.Memory.Shards <= 0 {
+		return nil, fmt.Errorf("expected >=1 shards, found: %v", conf.Memory.Shards)
+	}
+	if conf.Memory.Shards == 1 {
+		m.shards = []*shard{
+			{
+				items: map[string]item{},
+				ttl:   time.Second * time.Duration(conf.Memory.TTL),
+
+				compInterval:   interval,
+				lastCompaction: time.Now(),
+
+				mKeys:        stats.GetGauge("keys"),
+				mCompactions: stats.GetCounter("compaction"),
+			},
+		}
+	} else {
+		for i := 0; i < conf.Memory.Shards; i++ {
+			m.shards = append(m.shards, &shard{
+				items: map[string]item{},
+				ttl:   time.Second * time.Duration(conf.Memory.TTL),
+
+				compInterval:   interval,
+				lastCompaction: time.Now(),
+
+				mKeys:        stats.GetGauge(fmt.Sprintf("shard.%v.keys", i)),
+				mCompactions: stats.GetCounter(fmt.Sprintf("shard.%v.compaction", i)),
+			})
+		}
+	}
+
 	for k, v := range conf.Memory.InitValues {
-		items[k] = item{
+		m.getShard(k).items[k] = item{
 			value: []byte(v),
 			ts:    time.Time{},
 		}
 	}
-	return &Memory{
-		items:          items,
-		ttl:            time.Second * time.Duration(conf.Memory.TTL),
-		compInterval:   interval,
-		lastCompaction: time.Now(),
-		stats:          stats,
-		mCompactions:   stats.GetCounter("compaction"),
-		mKeys:          stats.GetGauge("keys"),
-	}, nil
+
+	return m, nil
 }
 
 //------------------------------------------------------------------------------
 
-func (m *Memory) compaction() {
-	if m.compInterval == 0 {
-		return
+func (m *Memory) getShard(key string) *shard {
+	if len(m.shards) == 1 {
+		return m.shards[0]
 	}
-	if time.Since(m.lastCompaction) < m.compInterval {
-		return
-	}
-	m.mCompactions.Incr(1)
-	for k, v := range m.items {
-		if v.ts.IsZero() {
-			continue
-		}
-		if time.Since(v.ts) >= m.ttl {
-			delete(m.items, k)
-		}
-	}
-	m.lastCompaction = time.Now()
+	h := xxhash.New64()
+	h.WriteString(key)
+	return m.shards[h.Sum64()%uint64(len(m.shards))]
 }
 
 // Get attempts to locate and return a cached value by its key, returns an error
 // if the key does not exist.
 func (m *Memory) Get(key string) ([]byte, error) {
-	m.RLock()
-	k, exists := m.items[key]
-	m.RUnlock()
+	shard := m.getShard(key)
+	shard.RLock()
+	k, exists := shard.items[key]
+	shard.RUnlock()
 	if !exists {
 		return nil, types.ErrKeyNotFound
 	}
@@ -154,49 +199,50 @@ func (m *Memory) Get(key string) ([]byte, error) {
 
 // Set attempts to set the value of a key.
 func (m *Memory) Set(key string, value []byte) error {
-	m.Lock()
-	m.compaction()
-	m.items[key] = item{value: value, ts: time.Now()}
-	m.mKeys.Set(int64(len(m.items)))
-	m.Unlock()
+	shard := m.getShard(key)
+	shard.Lock()
+	shard.compaction()
+	shard.items[key] = item{value: value, ts: time.Now()}
+	shard.mKeys.Set(int64(len(shard.items)))
+	shard.Unlock()
 	return nil
 }
 
 // SetMulti attempts to set the value of multiple keys, returns an error if any
 // keys fail.
 func (m *Memory) SetMulti(items map[string][]byte) error {
-	m.Lock()
-	m.compaction()
 	for k, v := range items {
-		m.items[k] = item{value: v, ts: time.Now()}
+		if err := m.Set(k, v); err != nil {
+			return err
+		}
 	}
-	m.mKeys.Set(int64(len(m.items)))
-	m.Unlock()
 	return nil
 }
 
 // Add attempts to set the value of a key only if the key does not already exist
 // and returns an error if the key already exists.
 func (m *Memory) Add(key string, value []byte) error {
-	m.Lock()
-	if _, exists := m.items[key]; exists {
-		m.Unlock()
+	shard := m.getShard(key)
+	shard.Lock()
+	if _, exists := shard.items[key]; exists {
+		shard.Unlock()
 		return types.ErrKeyAlreadyExists
 	}
-	m.compaction()
-	m.items[key] = item{value: value, ts: time.Now()}
-	m.mKeys.Set(int64(len(m.items)))
-	m.Unlock()
+	shard.compaction()
+	shard.items[key] = item{value: value, ts: time.Now()}
+	shard.mKeys.Set(int64(len(shard.items)))
+	shard.Unlock()
 	return nil
 }
 
 // Delete attempts to remove a key.
 func (m *Memory) Delete(key string) error {
-	m.Lock()
-	m.compaction()
-	delete(m.items, key)
-	m.mKeys.Set(int64(len(m.items)))
-	m.Unlock()
+	shard := m.getShard(key)
+	shard.Lock()
+	shard.compaction()
+	delete(shard.items, key)
+	shard.mKeys.Set(int64(len(shard.items)))
+	shard.Unlock()
 	return nil
 }